+++ R/KSJPrefCodes.R
@@ -1,9 +1,0 @@
-#' Corresponding Table Of Names And Codes Of Prefectures
-#'
-#' @format A data.frame with prefecture names and codes
-#'
-#' @source \url{http://nlftp.mlit.go.jp/ksj/gml/codelist/PrefCd.html}
-"KSJPrefCodes"
-
-# Suppres notes with R CMD check; see http://stackoverflow.com/a/12429344/5397672
-globalVariables("KSJPrefCodes")