+++ R/rest_in_peace.R
@@ -1,0 +1,21 @@
+#' Rest in peace, kokudo suuchi API...
+#'
+#' @rdname RIP_API
+#' @param ... Ignored.
+#'
+#' @export
+getKSJSummary <- function(...) {
+  .Defunct()
+}
+
+#' @rdname RIP_API
+#' @export
+getKSJURL <- function(...) {
+  .Defunct()
+}
+
+#' @rdname RIP_API
+#' @export
+choose_prefecture_code <- function(...) {
+  .Defunct()
+}