+++ R/zzz.R
@@ -1,3 +1,0 @@
-.onAttach <- function(...) {
-  packageStartupMessage(paste(strwrap(KSJcredit), collapse = "\n"))
-}