+++ NEWS.md
@@ -1,4 +1,11 @@
-# kokudosuuchi 0.4.2.9000
+# kokudosuuchi 1.0.0
+
+## Breaking change
+
+* kokudosuuchi drops all functionality related to the 'Kokudo Suuchi' API as the
+  API is no longer provided. Instead, a new function `read_ksj_data()` is
+  implemented to make it ease to read various geospatial data provided by
+  'Kokudo Suuchi'.
 
 # kokudosuuchi 0.4.2